requirements.txt
pyTelegramBotAPI~=4.2.2
PyYAML~=6.0
prometheus-client~=0.14.1
requests~=2.27.1
urllib3~=1.26.8
beautifulsoup4==4.10.0
fuzzywuzzy
pymorphy2